backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
pg8000==1.30.4
Pillow==10.0.1
PyYAML==6.0.1
gunicorn==21.2.0
python-decouple==3.8